addins {golem}	R Documentation

{golem} addins

Description

insert_ns() takes a selected character vector and wrap it in ns() The series of go_to_*() addins help you go to common files used in developing a {golem} application.

Usage


insert_ns()

go_to_start()

go_to_dev()

go_to_deploy()

go_to_run_dev()

go_to_app_ui()

go_to_app_server()

go_to_run_app()

[Package golem version 0.4.1 Index]