get_sysreqs {golem}	R Documentation

Get system requirements (Deprecated)

Description

This function retrieves information about the system requirements using the https://sysreqs.r-hub.io API. This function is now deprecated, and was moved to dockerfiler.

Usage


get_sysreqs(packages, quiet = TRUE, batch_n = 30)

Arguments

packages
character vector. Packages names.

quiet
Boolean. If TRUE the function is quiet.

batch_n
numeric. Number of simultaneous packages to ask.

Value

A vector of system requirements.

[Package golem version 0.4.1 Index]