app_prod {golem}	R Documentation

Is the app in dev mode or prod mode?

Description

Is the app in dev mode or prod mode?

Usage


app_prod()

app_dev()

Value

TRUE or FALSE depending on the status of getOption( "golem.app.prod")

[Package golem version 0.4.1 Index]