add_positconnect_file {golem}	R Documentation

Add an app.R at the root of your package to deploy on RStudio Connect

Description

Additionally, adds a .rscignore at the root of the {golem} project if the rsconnect package version is >= 0.8.25.

Usage

add_positconnect_file(pkg = get_golem_wd(), open = TRUE)

add_rstudioconnect_file(pkg = get_golem_wd(), open = TRUE)

add_shinyappsio_file(pkg = get_golem_wd(), open = TRUE)

add_shinyserver_file(pkg = get_golem_wd(), open = TRUE)

add_rscignore_file(pkg = get_golem_wd(), open = TRUE)

Arguments

pkg
Path to the root of the package. Default is get_golem_wd().

open
Should the created file be opened?

Value

Side-effect functions for file creation returning the path to the file, invisibly.

List of excluded files in .rscignore

.here

CODE_OF_CONDUCT.md

LICENSE{.md}

LICENCE{.md}

NEWS{.md}

README{.md,.Rmd,.HTML}

dev

man

tests

vignettes

Note

In previous versions, this function was called add_rconnect_file.

add_rstudioconnect_file is now deprecated; replace by add_positconnect_file().

Examples

# Add a file for Connect
if (interactive()) {
  add_positconnect_file()
}
# Add a file for Shiny Server
if (interactive()) {
  add_shinyserver_file()
}
# Add a file for Shinyapps.io
if (interactive()) {
  add_shinyappsio_file()
}

[Package golem version 0.5.1 Index]