js_handler_template {golem}	R Documentation

Golem's default custom templates

Description

These functions do not aim at being called as is by users, but to be passed as an argument to the add_js_handler() function.

Usage


js_handler_template(path, name = "fun", code = " ")

js_template(path, code = " ")

css_template(path, code = " ")

sass_template(path, code = " ")

Arguments

path
The path to the JS script where this template will be written.

name
Shiny's custom handler name.

code
JavaScript code to be written in the function.

Value

Used for side effect

See Also

add_js_handler()

[Package golem version 0.4.1 Index]