Create a Dockerfile for Shiny App

Build a container containing your Shiny App. add_dockerfile() creates a "classical" Dockerfile, while add_dockerfile_shinyproxy() and add_dockerfile_heroku() creates platform specific Dockerfile.

add_dockerfile(
  path = "DESCRIPTION",
  output = "Dockerfile",
  pkg = get_golem_wd(),
  from = paste0("rocker/r-ver:", R.Version()$major, ".", R.Version()$minor),
  as = NULL,
  port = 80,
  host = "0.0.0.0",
  sysreqs = TRUE,
  repos = c(CRAN = "https://cran.rstudio.com/"),
  expand = FALSE,
  open = TRUE,
  update_tar_gz = TRUE,
  build_golem_from_source = TRUE,
  extra_sysreqs = NULL
)

add_dockerfile_shinyproxy(
  path = "DESCRIPTION",
  output = "Dockerfile",
  pkg = get_golem_wd(),
  from = paste0("rocker/r-ver:", R.Version()$major, ".", R.Version()$minor),
  as = NULL,
  sysreqs = TRUE,
  repos = c(CRAN = "https://cran.rstudio.com/"),
  expand = FALSE,
  open = TRUE,
  update_tar_gz = TRUE,
  build_golem_from_source = TRUE,
  extra_sysreqs = NULL
)

add_dockerfile_heroku(
  path = "DESCRIPTION",
  output = "Dockerfile",
  pkg = get_golem_wd(),
  from = paste0("rocker/r-ver:", R.Version()$major, ".", R.Version()$minor),
  as = NULL,
  sysreqs = TRUE,
  repos = c(CRAN = "https://cran.rstudio.com/"),
  expand = FALSE,
  open = TRUE,
  update_tar_gz = TRUE,
  build_golem_from_source = TRUE,
  extra_sysreqs = NULL
)

Arguments

path
path to the DESCRIPTION file to use as an input.

output
name of the Dockerfile output.

pkg
Path to the root of the package. Default is get_golem_wd().

from
The FROM of the Dockerfile. Default is FROM rocker/r-ver: with R.Version()$major and R.Version()$minor.

as
The AS of the Dockerfile. Default it NULL.

port
The options('shiny.port') on which to run the Shiny App. Default is 80.

host
The options('shiny.host') on which to run the Shiny App. Default is 0.0.0.0.

sysreqs
boolean to check the system requirements

repos
character vector, the base URL of the repositories

expand
boolean, if TRUE each system requirement will be known his own RUN line

open
boolean, default is TRUE open the Dockerfile file

update_tar_gz
boolean, if TRUE and build_golem_from_source is also TRUE an updated tar.gz Package is created

build_golem_from_source
boolean, if TRUE no tar.gz Package is created and the Dockerfile directly mount the source folder to build it

extra_sysreqs
extra debian system requirements as character vector. Will be installed with apt-get install

Examples

# \donttest{
# Add a standard Dockerfile
if (interactive()){
   add_dockerfile()
}
# Add a Dockerfile for ShinyProxy
if (interactive()){
    add_dockerfile_shinyproxy()
}
# Add a Dockerfile for Heroku
if (interactive()){
    add_dockerfile_heroku()
}
# }