Regularized Generalized Linear Models

This functions fits a generalized linear model (GLM) using efficient optimization routines suitable to big data problems.

golem(family = c("gaussian", "binomial"), penalty = c("slope",
  "group_slope", "lasso"), solver = "fista", intercept = TRUE,
  standardize_features = TRUE, orthogonalize = TRUE, sigma = NULL,
  lambda = NULL, fdr = 0.2, n_lambda = 100,
  lambda_min_ratio = NULL, tol_rel_gap = 1e-06, tol_infeas = 1e-06,
  max_passes = 10000, diagnostics = FALSE)

Arguments

family
response type. See Families for details.

penalty
the regularization penalty to use. See Penalties for details.

solver
the numerical solver to use. See Solvers for details.

intercept
whether to fit an intercept

standardize_features
whether to standardize features (predictors)

orthogonalize
whether x should be orthogonalized. Note that setting this to TRUE when x is sparse will through an error. (only applies to Group SLOPE)

sigma
noise estimate (only applies to SLOPE and Group SLOPE)

lambda
either a character vector indicating the method used to construct the lambda path or

fdr
target false discovery rate (only applies to SLOPE and Group SLOPE)

n_lambda
length of regularization path (only relevant for lasso)

lambda_min_ratio
smallest value for lambda as a fraction of \(\lambda_\mathrm{max}\) (only applies to lasso)

tol_rel_gap
relative tolerance threshold for duality gap check

tol_infeas
tolerance threshold for infeasibility

max_passes
maximum number of passes for optimizer

diagnostics
should diagnostics be saved for the model fit (timings, primal and dual objectives, and infeasibility)

Value

An object of class "Golem".

Details

The objective for each model is simply the loss function for each family plus a penalty term.

Methods

fit(x, y, groups = NULL, warm_start = TRUE, ...)

This method fits models specified by golem().

x

the feature matrix, which can be either a dense matrix of the standard matrix class, or a sparse matrix inheriting from Matrix::sparseMatrix Data frames will be converted to matrices internally.

y

the response. For Gaussian models, this must be numeric; for binomial models, it can be a factor.

groups

a vector of integers giving the group membership of each feature (only applies to Group SLOPE)

warm_start

whether to use the coefficient estimates from the previous fit when refitting the model using new data. Provided that the same penalty (with approximately the same parameters) id used, setting this to true might lead to substantial performance boosts.

…

arguments that will be used to modify the original model specification from the call to golem(). Note that arguments pushed through this interface will modify the model, which might have unintended consequences.

coef()

Return the coefficients from the model fit (after dropping extraneous dimensions). If you prefer to always return a three-dimensional array, call model$coefficients instead.

predict(x, type = c("link", "response", "class"))

Return predictions from models fit by golem() based on new data.

x

new data to make predictions for.

type

type of predictions to make. "link" gives the linear predictors, "response" gives the result of applying the link function, and "class" gives class predictions.

Families

Gaussian

The Gaussian model (Ordinary Least Squares) minimizes the following objective.

$$ ||\boldsymbol{y} - \boldsymbol{X\beta}||_2^2 $$

Binomial

The binomial model (logistic regression) has the following objective.

$$ \sum_{i=1}^n \log\left[1+ \exp\left(- y_i\boldsymbol{x}_i'\boldsymbol{\beta} \right) \right] $$

Penalties

Models fit by golem::golem() can be regularized via several penalties.

Lasso

The Lasso penalizes coefficients using the L1 norm. The penalty term in the lagrangian form of the loss function is

$$ \lambda \sum_{j=1}^p |\beta_j| $$

SLOPE

SLOPE (Sorted L-One Penalized Estimation) is an extension of the Lasso. Unlike the latter, however, SLOPE uses a non-increasing sequence of \(\lambda\)---one for each coefficient. The penalty term looks like

$$ \sigma \sum_{i=j}^p \lambda_j |\beta|_{(j)} $$

Group SLOPE

Group SLOPE is an extension of Group LASSO. It applies the following penalty

$$ J_\lambda(\beta) = \sum_{j=1}^p \lambda_j |W||\beta||_{I,X}|_{(j)} $$

Solvers

There is currently a single solver available for golem::golem.

FISTA

FISTA (Fast Iterative Shrinking-Tresholding Algorithm) is an extension of the classical gradient algorithm.

Examples

# Gaussian response, slope penalty (default) --------------------------------

# Specify the model
gaussian_model <- golem(family = "gaussian")

# Fit the model
gaussian_model$fit(abalone$x, abalone$y)

# Get coefficients from the model fit
gaussian_model$coef()

#>    (Intercept)            sex         infant         length       diameter 
#>      4.7343417      0.0000000     -0.8427958      0.0000000      7.1484161 
#>         height   weight_whole weight_shucked weight_viscera   weight_shell 
#>     10.8508506      0.0000000     -8.0753268      0.0000000     16.5077169 

# Binomial response, lasso penalty ------------------------------------------

binomial_model <- golem(family = "binomial", penalty = "lasso")

binomial_model$fit(heart$x, heart$y)