add_rstudioconnect_file {golem}	R Documentation

Add an app.R at the root of your package to deploy on RStudio Connect

Description

Add an app.R at the root of your package to deploy on RStudio Connect

Usage


add_rstudioconnect_file(pkg = get_golem_wd(), open = TRUE)

add_shinyappsio_file(pkg = get_golem_wd(), open = TRUE)

add_shinyserver_file(pkg = get_golem_wd(), open = TRUE)

Arguments

pkg
Path to the root of the package. Default is get_golem_wd().

open
Should the created file be opened?

Value

The path to the file, invisibly.

Note

In previous versions, this function was called add_rconnect_file.

Examples


# Add a file for Connect
if (interactive()) {
  add_rstudioconnect_file()
}
# Add a file for Shiny Server
if (interactive()) {
  add_shinyserver_file()
}
# Add a file for Shinyapps.io
if (interactive()) {
  add_shinyappsio_file()
}

[Package golem version 0.4.1 Index]