expect_shinytag {golem}	R Documentation

Test helpers

Description

These functions are designed to be used inside the tests in your Shiny app package.

Usage


expect_shinytag(object)

expect_shinytaglist(object)

expect_html_equal(ui, html, ...)

expect_running(sleep, R_path = NULL)

Arguments

object
the object to test

ui
output of an UI function

html
deprecated

...
arguments passed to testthat::expect_snapshot()

sleep
number of seconds

R_path
path to R. If NULL, the function will try to guess where R is.

Value

A testthat result.

[Package golem version 0.4.1 Index]