use_recommended_deps {golem}	R Documentation

Add recommended elements

Description

use_recommended_deps

Adds shiny, DT, attempt, glue, golem, htmltools to dependencies

use_recommended_tests

Adds a test folder and copy the golem tests

Usage


use_recommended_deps(
  pkg = get_golem_wd(),
  recommended = c("shiny", "DT", "attempt", "glue", "htmltools", "golem")
)

use_recommended_tests(
  pkg = get_golem_wd(),
  spellcheck = TRUE,
  vignettes = TRUE,
  lang = "en-US",
  error = FALSE
)

Arguments

pkg
Path to the root of the package. Default is get_golem_wd().

recommended
A vector of recommended packages.

spellcheck
Whether or not to use a spellcheck test.

vignettes
Logical, TRUE to spell check all rmd and rnw files in the vignettes/ folder.

lang
Preferred spelling language. Usually either "en-US" or "en-GB".

error
Logical, indicating whether the unit test should fail if spelling errors are found. Defaults to FALSE, which does not error, but prints potential spelling errors

Value

Used for side-effects.

[Package golem version 0.4.1 Index]